backend/requirements.txt
fastapi
httpx
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[argon2]
argon2-cffi
google-generativeai
apscheduler